requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
PyPDF2==3.0.1
python-docx==1.1.0

gunicorn==21.2.0
